Crate tauri_plugin_dialog

Native system dialogs for opening and saving files along with message dialogs.

Structs

Dialog

Access to the dialog APIs.

FileDialogBuilder

The file dialog builder.

MessageDialogBuilder

A builder for message dialogs.

Enums

Error

FileAccessMode

FilePath

Represents either a filesystem path or a URI pointing to a file such as file:// URIs or Android content:// URIs.

MessageDialogButtons

Set of button that will be displayed on the dialog

MessageDialogKind

Types of message, ask and confirm dialogs.

MessageDialogResult

Result of a message dialog

PickerMode

Traits

DialogExt

Extensions to [tauri::App], [tauri::AppHandle], [tauri::WebviewWindow], [tauri::Webview] and [tauri::Window] to access the dialog APIs.

Functions

init

Initializes the plugin.

Type Aliases

Result